server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.0",
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "celery>=5.5.3",
    "cloudinary>=1.44.1",
    "crewai>=0.203.1",
    "fastapi[standard]>=0.118.3",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "pinecone[asyncio]>=7.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "redis[hiredis]>=6.4.0",
    "sqlalchemy>=2.0.44",
    "unstructured[pdf]>=0.18.15",
    "uvicorn[standard]>=0.37.0",
]
